requirements.txt
certifi==2017.11.5
chardet==3.0.4
dominate==2.3.1
idna==2.6
numpy==1.21.0
Pillow==9.0.1
PyYAML==5.4
pyzmq==16.0.3
requests==2.20.1
six==1.11.0
torch==0.4.0
torchfile==0.1.0
torchvision==0.2.1
tornado==4.5.3
urllib3==1.26.5
visdom==0.1.7
